+++ src/temperature/solvers/alt_dir.py
@@ -1,0 +1,232 @@
+import numba
+import numpy as np
+from numpy.typing import NDArray
+
+from src.boundary_conditions import BoundaryConditionType
+import src.parameters as cfg
+from src.temperature.coefficient_smoothing.coefficients import c_smoothed, k_smoothed
+from src.temperature.coefficient_smoothing.delta import get_max_delta
+from src.temperature.solvers.base import HeatTransferSolver
+from src.utils import solve_tridiagonal
+
+
+class AltDirSolver(HeatTransferSolver):
+    @staticmethod
+    @numba.jit(nopython=True)
+    def _compute_sweep_x(
+        u: NDArray[np.float64],
+        iter_u: NDArray[np.float64],
+        temp_u: NDArray[np.float64],
+        alpha: NDArray[np.float64],
+        beta: NDArray[np.float64],
+        dx: float,
+        dy: float,
+        dt: float,
+        right_cond_type: int,
+        left_cond_type: int,
+        delta: float,
+        time: float = 0.0,
+    ) -> NDArray[np.float64]:
+        n_y, n_x = u.shape
+        inv_dx2 = 1.0 / (dx * dx)
+        inv_dy2 = 1.0 / (dy * dy)
+
+        lbc = bc.get_left_bc_1(time, n_y)
+        rbc = bc.get_right_bc_1(time, n_y)
+
+        for j in range(1, n_y - 1):
+            if left_cond_type == cfg.DIRICHLET:
+                alpha[0] = 0.0
+                beta[0] = lbc[j]
+            else:
+                alpha[0] = 1.0
+                beta[0] = 0.0
+            for i in range(1, n_x - 1):
+                inv_c = 1.0 / c_smoothed(u[j, i], delta)
+                a_i = (
+                    -dt
+                    * 0.5
+                    * k_smoothed(0.5 * (u[j, i + 1] + u[j, i]), delta)
+                    * inv_c
+                    * inv_dx2
+                )
+                b_i = (
+                    1.0
+                    + dt
+                    * (
+                        k_smoothed(0.5 * (u[j, i + 1] + u[j, i]), delta)
+                        + k_smoothed(0.5 * (u[j, i] + u[j, i - 1]), delta)
+                    )
+                    * inv_c
+                    * inv_dx2
+                    * 0.5
+                )
+                c_i = (
+                    -dt
+                    * 0.5
+                    * k_smoothed(0.5 * (u[j, i] + u[j, i - 1]), delta)
+                    * inv_c
+                    * inv_dx2
+                )
+
+                rhs_i = u[j, i] + dt * 0.5 * inv_c * inv_dy2 * (
+                    k_smoothed(0.5 * (u[j + 1, i] + u[j, i]), delta)
+                    * (u[j + 1, i] - u[j, i])
+                    - k_smoothed(0.5 * (u[j, i] + u[j - 1, i]), delta)
+                    * (u[j, i] - u[j - 1, i])
+                )
+
+                alpha[i] = -a_i / (b_i + c_i * alpha[i - 1])
+                beta[i] = (rhs_i - c_i * beta[i - 1]) / (b_i + c_i * alpha[i - 1])
+
+            temp_u[j, n_x - 1] = (
+                rbc[j]
+                if right_cond_type == 1
+                else beta[n_x - 2] / (1.0 - alpha[n_x - 2])  # Neumann
+            )
+
+            for i in range(n_x - 2, -1, -1):
+                temp_u[j, i] = alpha[i] * temp_u[j, i + 1] + beta[i]
+
+        temp_u[0, :] = u[0, :]
+        temp_u[n_y - 1, :] = u[n_y - 1, :]
+
+        return temp_u
+
+    @staticmethod
+    @numba.jit(nopython=True)
+    def _compute_sweep_y(
+        temp_u: NDArray[np.float64],
+        iter_u: NDArray[np.float64],
+        new_u: NDArray[np.float64],
+        alpha: NDArray[np.float64],
+        beta: NDArray[np.float64],
+        dx: float,
+        dy: float,
+        dt: float,
+        top_cond_type: int,
+        bottom_cond_type: int,
+        delta: float,
+        time: float = 0.0,
+    ) -> NDArray[np.float64]:
+        n_y, n_x = temp_u.shape
+        inv_dx2 = 1.0 / (dx * dx)
+        inv_dy2 = 1.0 / (dy * dy)
+
+        bbc = bc.get_bottom_bc_1(time, n_x)
+        tbc = bc.get_top_bc_1(time, n_x)
+        phi = bc.get_top_bc_2(time)
+        psi, ksi = bc.get_top_bc_3(time)
+
+        for i in range(1, n_x - 1):
+            if bottom_cond_type == cfg.DIRICHLET:
+                alpha[0] = 0.0
+                beta[0] = bbc[i]
+            else:  # Neumann
+                alpha[0] = 1.0
+                beta[0] = 0.0
+            for j in range(1, n_y - 1):
+                inv_c = 1.0 / c_smoothed(temp_u[j, i], delta)
+                a_j = (
+                    -dt
+                    * 0.5
+                    * k_smoothed(0.5 * (temp_u[j + 1, i] + temp_u[j, i]), delta)
+                    * inv_c
+                    * inv_dy2
+                )
+                b_j = (
+                    1.0
+                    + dt
+                    * (
+                        k_smoothed(0.5 * (temp_u[j + 1, i] + temp_u[j, i]), delta)
+                        + k_smoothed(0.5 * (temp_u[j, i] + temp_u[j - 1, i]), delta)
+                    )
+                    * inv_c
+                    * inv_dy2
+                    * 0.5
+                )
+                c_j = (
+                    -dt
+                    * 0.5
+                    * k_smoothed(0.5 * (temp_u[j, i] + temp_u[j - 1, i]), delta)
+                    * inv_c
+                    * inv_dy2
+                )
+
+                rhs_j = temp_u[j, i] + dt * 0.5 * inv_c * inv_dx2 * (
+                    k_smoothed(0.5 * (temp_u[j, i + 1] + temp_u[j, i]), delta)
+                    * (temp_u[j, i + 1] - temp_u[j, i])
+                    - k_smoothed(0.5 * (temp_u[j, i] + temp_u[j, i - 1]), delta)
+                    * (temp_u[j, i] - temp_u[j, i - 1])
+                )
+
+                alpha[j] = -a_j / (b_j + c_j * alpha[j - 1])
+                beta[j] = (rhs_j - c_j * beta[j - 1]) / (b_j + c_j * alpha[j - 1])
+
+            if top_cond_type == cfg.DIRICHLET:
+                new_u[n_y - 1, i] = tbc[i]
+            elif top_cond_type == cfg.NEUMANN:
+                new_u[n_y - 1, i] = (dy * phi + beta[n_y - 2]) / (1.0 - alpha[n_y - 2])
+            else:  # ROBIN
+                new_u[n_y - 1, i] = (dy * psi + beta[n_y - 2]) / (
+                    1 - alpha[n_y - 2] - dy * ksi
+                )
+
+            # Вычисление температуры на новом временном слое
+            for j in range(n_y - 2, -1, -1):
+                new_u[j, i] = alpha[j] * new_u[j + 1, i] + beta[j]
+
+        new_u[:, 0] = temp_u[:, 0]
+        new_u[:, n_x - 1] = temp_u[:, n_x - 1]
+
+        return new_u
+
+    def solve(
+        self,
+        u: NDArray[np.float64],
+        sf: NDArray[np.float64],
+        time: float = 0.0,
+        iters: int = 1,
+    ) -> NDArray[np.float64]:
+
+        self._iter_u = np.copy(u)
+
+        # Run the x-direction sweep iterations
+        for i in range(iters):
+            delta = cfg.delta if self.fixed_delta else get_max_delta(self._iter_u)
+            self._temp_u = self._compute_sweep_x(
+                u=u,
+                iter_u=self._iter_u,
+                temp_u=self._temp_u,
+                alpha=self._alpha,
+                beta=self._beta,
+                dx=self.geometry.dx,
+                dy=self.geometry.dy,
+                dt=self.geometry.dt,
+                right_cond_type=self.right_cond_type,
+                left_cond_type=self.left_cond_type,
+                delta=delta,
+                time=time,
+            )
+            self._iter_u = np.copy(self._temp_u)
+
+        # Run the y-direction sweep iterations
+        for i in range(iters):
+            delta = cfg.delta if self.fixed_delta else get_max_delta(self._temp_u)
+            self._new_u = self._compute_sweep_y(
+                temp_u=self._temp_u,
+                iter_u=self._iter_u,
+                new_u=self._new_u,
+                alpha=self._alpha,
+                beta=self._beta,
+                dx=self.geometry.dx,
+                dy=self.geometry.dy,
+                dt=self.geometry.dt,
+                top_cond_type=self.top_cond_type,
+                bottom_cond_type=self.bottom_cond_type,
+                delta=delta,
+                time=time,
+            )
+            self._iter_u = np.copy(self._new_u)
+
+        return self._new_u
